[Chore #107707438] Codebase clean up, moved testing scripts to appropriate test files
