+++ codango/userprofile/tests/test_views.py
@@ -1,36 +1,39 @@
-from django.test import Client
-from django.contrib.auth.models import User
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from selenium import webdriver
 from selenium.webdriver.common.keys import Keys
 
-from django.test.utils import setup_test_environment
-
-setup_test_environment()
-
 
 class UserProfileTest(StaticLiveServerTestCase):
+    fixtures = ['users.json']
 
     def setUp(self):
-        self.client = Client()
-        self.user = User.objects.create(username='jubril', password='issa')
-        self.user.set_password('shuaib')
-        self.user.save()
-        self.login = self.client.login(username='Abiodun', password='shuaib')
+        self.browser = webdriver.PhantomJS()
+        self.browser.set_window_size(1400, 1000)
+        self.browser.implicitly_wait(10)
 
+    def tearDown(self):
+        self.browser.quit()
 
-class FollowUserProfileTest(StaticLiveServerTestCase):
+    def test_can_reach_profile_page(self):
+        self.browser.get(self.live_server_url)
 
-    def setUp(self):
-        self.client = Client()
-        self.user1 = User.objects.create_user(
-            username='golden', password='abiodun')
-        self.user2 = User.objects.create_user(
-            username='jubril', password='issa')
-        self.user1.save()
-        self.user2.save()
-        self.login = self.client.login(username='golden', password='abiodun')
+        body = self.browser.find_element_by_tag_name('body')
+        self.assertIn('Codango', body.text)
 
-    def test_a_logged_in_user_can_follow_a_registered_user(self):
-        response = self.client.post('/user/golden/follow')
-        self.assertEqual(response.status_code, 200)
+        # logging in username and password
+        username_field = self.browser.find_element_by_name('username')
+        username_field.send_keys('lade')
+
+        password_field = self.browser.find_element_by_name('password')
+        password_field.send_keys('password')
+        password_field.send_keys(Keys.RETURN)
+
+        # username and password accepted
+        body = self.browser.find_element_by_tag_name('body')
+        self.assertIn('Share', body.text)
+
+        # User profile
+        self.browser.find_element_by_link_text('lade').click()
+        self.browser.find_element_by_link_text('View Profile').click()
+        body = self.browser.find_element_by_tag_name('body')
+        self.assertIn('@lade', body.text)
